Changed if statement for the behavior variable
Changed the code so it would reflect the material used to arrive at the latest results of experiment four. Added the line where the agent uses full Kelly when its current wealth is equal to initial wealth. Prior versions of this code had the agent using a fraction of Kelly when this held true.